NCI-Agency/anet#2926: Remove try-out assessment results components
Decided on which layout to use for the assessment results and
therefore removed the components for the layout not to be used.
